article-api: Make use of shared scalatra code
